@@ -2,22 +2,23 @@
 
 namespace App\GraphQL\Fields;
 
-use GraphQL\Type\Definition\Type;
-use Folklore\GraphQL\Support\Field;
 use Folklore\GraphQL\Support\Facades\GraphQL;
+use Folklore\GraphQL\Support\Field;
+use GraphQL\Type\Definition\Type;
 
-
-class MediasField extends Field {
-    
+class MediasField extends Field
+{
     protected $attributes = [
-		'description' => 'A picture'
+        'description' => 'A picture',
     ];
 
-    public function type(){
+    public function type()
+    {
         return Type::listOf(GraphQL::type('Media'));
     }
     
-    public function resolve($root, $args){
+    public function resolve($root, $args)
+    {
         return $root->medias;
     }
 }
