@@ -2,26 +2,28 @@
 
 namespace App\GraphQL\Query;
 
-use GraphQL\Type\Definition\Type;
-use Folklore\GraphQL\Support\Query;
-use Folklore\GraphQL\Support\Facades\GraphQL;
 use App\Models\Category;
+use Folklore\GraphQL\Support\Facades\GraphQL;
+use Folklore\GraphQL\Support\Query;
+use GraphQL\Type\Definition\Type;
 
-class CategoriesQuery extends Query{
-    
+class CategoriesQuery extends Query
+{
     protected $attributes = [
-        'name' => 'categories'
+        'name' => 'categories',
     ];
 
-    public function type(){
+    public function type()
+    {
         return Type::listOf(GraphQL::type('Category'));
     }
 
-    public function args(){
-
+    public function args()
+    {
     }
 
-    public function resolve($root, $args){
+    public function resolve($root, $args)
+    {
         return Category::with('products')->get();
     }
 }
