@@ -2,21 +2,20 @@
 
 namespace App\GraphQL\Type;
 
-use GraphQL\Type\Definition\Type;
-use Folklore\GraphQL\Support\Type as BaseType;
 use App\GraphQL\Fields\CategoryProductField;
+use Folklore\GraphQL\Support\Type as BaseType;
+use GraphQL\Type\Definition\Type;
 
-class CategoryType extends BaseType{
-
+class CategoryType extends BaseType
+{
     public function fields()
     {
         return [
-            'name'=>[
-                'type'=>Type::string(),
-                'description'=>"Name of Product"
+            'name' => [
+                'type' => Type::string(),
+                'description' => "Name of Product",
             ],
-            'products'=>CategoryProductField::class
+            'products' => CategoryProductField::class,
         ];
     }
-
 }
