@@ -2,17 +2,15 @@
 
 namespace App\GraphQL\Type;
 
-use GraphQL\Type\Definition\Type;
-use Folklore\GraphQL\Support\Type as BaseType;
 use App\GraphQL\Fields\MediaUrlField;
+use Folklore\GraphQL\Support\Type as BaseType;
 
-class MediaType extends BaseType{
-    
-    public function fields(){
+class MediaType extends BaseType
+{
+    public function fields()
+    {
         return [
-            'url'=>MediaUrlField::class
+            'url' => MediaUrlField::class,
         ];
     }
-
-
 }
