@@ -2,15 +2,13 @@
 
 namespace App\Http\Controllers\Auth;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use Laravel\Socialite\Facades\Socialite;
 use App\User;
 use Auth;
+use Laravel\Socialite\Facades\Socialite;
 
 class SocialAuthController extends Controller
 {
-
     public function __construct()
     {
         $this->middleware('guest');
@@ -24,7 +22,7 @@ class SocialAuthController extends Controller
     protected $providers = [
         // 'github',
         'facebook',
-        'google'
+        'google',
         // 'twitter'
     ];
 
@@ -36,7 +34,7 @@ class SocialAuthController extends Controller
      */
     public function redirectToProvider($driver)
     {
-        if( ! $this->isProviderAllowed($driver) ) {
+        if (! $this->isProviderAllowed($driver)) {
             return $this->sendFailedResponse("{$driver} is not currently supported");
         }
 
@@ -54,7 +52,7 @@ class SocialAuthController extends Controller
      * @param $driver
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function handleProviderCallback( $driver )
+    public function handleProviderCallback($driver)
     {
         try {
             $user = Socialite::driver($driver)->user();
@@ -63,8 +61,8 @@ class SocialAuthController extends Controller
         }
 
         // check for email in returned user
-        if(empty( $user->email )){
-            $user->email=$user->id.'@aammui.com';
+        if (empty($user->email)) {
+            $user->email = $user->id.'@aammui.com';
         }
 
         return $this->loginOrCreateAccount($user, $driver);
@@ -98,12 +96,12 @@ class SocialAuthController extends Controller
         $user = User::where('email', $providerUser->getEmail())->first();
 
         // if user already found
-        if( $user ) {
+        if ($user) {
             // update the avatar and provider that might have changed
             $user->update([
                 'provider' => $driver,
                 'provider_id' => $providerUser->id,
-                'access_token' => $providerUser->token
+                'access_token' => $providerUser->token,
             ]);
         } else {
             // create a new user
@@ -114,7 +112,7 @@ class SocialAuthController extends Controller
                 'provider_id' => $providerUser->getId(),
                 'access_token' => $providerUser->token,
                 // user can use reset password to create a password
-                'password' => ''
+                'password' => '',
             ]);
         }
 
@@ -134,5 +132,4 @@ class SocialAuthController extends Controller
     {
         return in_array($driver, $this->providers) && config()->has("services.{$driver}");
     }
-    
 }
