@@ -2,14 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Product;
+use Illuminate\Http\Request;
 
 class SearchController extends Controller
 {
-    public function search(Request $request){
-        $products=Product::where('name','like','%'.$request->q.'%')
+    public function search(Request $request)
+    {
+        $products = Product::where('name', 'like', '%'.$request->q.'%')
             ->get();
-        return view('search.index',compact('products'));
+
+        return view('search.index', compact('products'));
     }
 }
