@@ -24,10 +24,10 @@ class ProductRequest extends FormRequest
     public function rules()
     {
         return [
-            'name'=>'required',
-            'condition'=>'required',
-            'negotiable'=>'required',
-            'price'=>'required'
+            'name' => 'required',
+            'condition' => 'required',
+            'negotiable' => 'required',
+            'price' => 'required',
         ];
     }
 }
