@@ -6,14 +6,14 @@ use Illuminate\Database\Eloquent\Model;
 
 class ShopProfile extends Model
 {
-    protected $dates =[
-        'established_at'
-    ];  
+    protected $dates = [
+        'established_at',
+    ];
 
-    protected $fillable=['info','address','opening_hours','sologon','user_id','established_at'];
+    protected $fillable = ['info','address','opening_hours','sologon','user_id','established_at'];
 
     // public function setEstablishedAtAttribute($value)
     // {
-    //     $this->attributes['established_at'] = 
+    //     $this->attributes['established_at'] =
     // }
 }
