@@ -1,13 +1,40 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-$factory->define(App\User::class, function (Faker $faker) {
-    $g=['women','men'];
-    return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'password' => '$2y$10$TKh8H1.PfQx37YgCzwiKb.KjNyWgaHb9cbcoQgdIVFlYg7B77UdFm', // secret
-        'remember_token' => str_random(10)
-    ];
-});
+use App\Domain\Users\Models\User;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * Class UserFactory
+ *
+ * @package Database\Factories
+ */
+class UserFactory extends Factory
+{
+    /**
+     * @var string
+     */
+    protected $model = User::class;
+
+    /**
+     * @return array
+     */
+    public function definition(): array
+    {
+        return [
+            'name'     => $this->faker->name,
+            'email'    => $this->faker->unique()->safeEmail,
+            'password' => bcrypt("password"),
+        ];
+    }
+
+    public function verified()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'email_verified_at' => now(),
+            ];
+        });
+    }
+}
