@@ -2,6 +2,8 @@
 
 namespace Database\Seeders;
 
+use App\Domain\Inventory\Models\Product;
+use Database\Factories\ProductFactory;
 use Illuminate\Database\Seeder;
 
 /**
@@ -18,7 +20,7 @@ class ProductSeeder extends Seeder
      */
     public function run()
     {
-        $products = factory('App\Models\Product', 20)->create();
+        $products = ProductFactory::new()->count(20)->create();
 
         foreach ($products as $item) {
             $medias = factory('App\Models\Media', rand(1, 5))->create();
